Put all whitespace to hidden channel in the grammar.
This makes the grammar considerably simpler and easier to write; parsing
errors are more useful as well.
The only drawback is that it requires extra hand-written code to
post-process the AST in order to determine whether an expression is a
list or concatenation of elemental expressions. This could be handled in
grammar description, ANTLRv3 provides means for it (validating semantic
predicates plus backtracking), but unfortunately, the Python backend is
a bit less tested and can't generate correct code in this case. We may
rewrite it properly later.