Change the grammar to use semicolons to terminate statements.
The language is reminiscent of C languages, so it won't be surprising.
This adds a bit of verbosity to simple oneliners, but is more convenient
for larger expressions (includedirs, defines, etc.) and these are
surprisingly common even in small projects.

Makes the grammar simpler, too, as it no longer needs to treat the
NEWLINE token differently based on context.