Add a pass for making variables from default properties values.
If a property isn't set or modified in the bakefile, then it's default
value is used. This was done on demand, i.e. when the property was first
used. Unfortunately, this bypassed all the processing passes, leading to
problematic code, failures or incorrect output.

By adding this step, all unset properties are added right after
splitting the model into toolset-specific parts. They are then processed
as usual, possibly optimized out etc.

Performance-wise, there shouldn't be much of an impact: the properties
exist because the toolset *uses* them for something, so their default
values will be evaluated sooner or later anyway.