Remove empty submodules.
If a submodule becomes empty after per-toolset processing of conditions
(i.e. some targets don't apply to this toolset) or the toolset is
disabled for this module, remove the empty module from the project. This
simplifies toolset's generator, because it doesn't have to worry about
unused parts of the model: there are none.