gnu-osx: Fix .d files creation with multiple -arch options.
Apple's GCC doesn't handle the standard -MD -MP flags (which are used to
generate .d files with object files' dependencies on sources and
headers) in presence of multiple -arch options. Clang can handle it, but
we must support GCC. So we run GCC's preprocessor once more to generate
the dependencies, but let's not do it unless necessary, because it a)
costs some time and b) may omit some deps.