Make iterating over ListExpr easier.
Implement support for standard Python iteration, so it's no longer
necessary to access the 'items' member directly and one can iterate over
the expression (or get its length or convert it to a list…) instead.

This works even with some other expression types (NullExpr and IfExpr)
to make it easy to write code that doesn't have to worry about a
condition being used in some place where the author didn't expect it.