Run antlr in the directory containing the grammar file.
Under Cygwin antlr outputs the files it creates in the current directory, not
the directory of the input file, as expected by the makefile. This looks like
an antlr bug but it's easier to work around it in our makefile by just running
antlr from the directory containing the grammar file instead of trying to fix
it in in antlr itself, so do this for now.