Use inheritance to match solution and project configs.
Solution has the union of all included projects' configurations.
If a project doesn't have some of the solution configurations, a
suitable substitute must be used instead.

Instead of basic this choice on substrings in configuration name (a
config called "Debug" is a good fit for sln-config "My Debug"), as
Visual Studio does, take advantage of the information that Bakefile has
about which configuration is based on which. If one configuration is
derived from the other, they are somehow related and one is a good
replacement fit for the other.

Finally, warn about poor substitutes.