Allow upgrading from previous version in the Windows installer.
Use MajorUpgrade element new in Wix 3.5 to do this, this is the simplest way
if the default behaviour is good enough which is the case for us. Notice that
we only disallow downgrades to avoid a bogus ICE61 Windows installer warning
when creating the setup, we could use "AllowDowngrades='yes'" and allow it
too.

Also don't hard code the same Product GUID for all versions but generate it
automatically for each new one instead.