Fix @srcdir normalization for good.
Instead of keeping track of the module in which the path is used and
normalizing based on that, take advantage of source code position
provided by ANTLR and available for, among other things,  every
expression. This makes the normalization 100% accurate and reliable,
provided that all path expressions have 'pos' attribute.