gnu: always set soname on shared libs.
It is an established best practice to always use sonames on Unix
platforms that have them, so that the runtime linker behaves predictably
and doesn't use unusual paths from link-time. This is useful even if the
shared library is unversioned, as is currently the case with bkl.

This fixes linking with dlls from another submodule, they now behave as
expected at runtime.