Don't use {...} around GUIDs.
Neither internally nor, more importantly, in vs*.guid properties. Not
only are the brackets not part of the usual (outside of the Microsoft
world, anyway) [GU]UID representation, they need to be quoted in
Bakefile for use in GUIDs, because they have another meaning in the
syntax. That would be too annoying, so use the bracket-less form.