Rename {exe,lib,dll}name properties to "basename".
There are several reasons for this renaming: it is consistent, allowing
easy manipulation from templates. It is a natural choice (just see the
documentation for the properties: they were all explained as "base name
of something"). Last but not least, "dllname" or "exename" don't make
much sense after the targets were renamed to shared-library and program
respectively.