MSVS: don't share intermediate directory between projects.
Visual Studio (especially 2010+ with its .tlog files) has issues with
building multiple projects in the same intermediate directory. Among the
problems exhibited by VS2010 and 2012 are cl.exe failures because a log
file is already open by another parallel process and rebuilds due to the
same object file appearing in multiple targets.

The safest route to avoid these problems is to simply not do that, so we
use e.g. Debug/$(id)/ intermediate directory instead of Debug as soon as
there's more than one project file in the same directory.