Handle srcdir setting on imported files correctly.
Filenames in imported files should be relative to the location of the
file too, not to the location of the file it is imported from. Likewise,
srcdir directive should be usable to change the default for the imported
file.

Previously, srcdir was set on modules in the model. With non-module bkl
files in the mix, it had to be reworked to be purely filename-based.