Introduce single-quoted strings as an alternative to double-quotes.
At the moment, they both behave identically, except for making it
possible to express the other kind of quote literal without escaping.

But the real purpose of this commit is preparation for changing the
semantics of double-quoted strings to be more like shell's quoting, i.e.
to interpret $ inside them. Single-quoted strings are then needed for
interpretation-less quoting.