Recognize path anchors like @srcdir at grammar level.
Previously, they were parsed when normalizing an untyped LiteralExpr
expression into a PathType, by manually parsing the string. This meant
that an anchored path string could be interpreted as a verbatim string
and not formatted correctly in the output.

Now that the anchors are part of the grammar, any explicit use of an
anchor will make the literal a PathExpr, with all the associated
goodness of  correct validation, formatting etc.