Fix ConfigurationProxy for linker properties.
ConfigurationProxy works for resolving conditions for a single target.
In some cases, however, values are obtained from multiple targets, e.g.
in NativeLinkedType.get_ldlibs() and others.

Provide a way to explicitly perform the proxying with apply_subst()
call. This is not particularly elegant (in that the code cannot use the
proxy transparently and must be aware of it in some special cases), but
it gets the job done.