Add bkl.api.CustomStep.
This extension interface allows hooking up into Bakefile processing at
various point (although only two points are implemented now).  For
example, it can be used to generate auxiliary files or to preprocess the
model in ways that the Bakefile language doesn't support or that would
be too inconvenient to do manually.