VS: Fix broken check for empty archs.
If the 'archs' property was set conditionally e.g. only for OS X, it
wouldn't be recognized as empty by VS toolset. This was due to variables
normalization that translated 'archs' to [] in this case. Fixed by
treating empty lists as nulls in this situation, which is The Right
Thing To Do anyway.