Define _DEBUG/NDEBUG for Windows RC files compilation in MSVS projects.
Although the native Visual Studio projects only define these symbols for C/C++
files and not RC ones by default, it's very convenient to have them defined
for the RC compilation too in practice and doing this is also more consistent
with our handling of "defines" which are used for both C/C++ and RC files, so
do define [one of] these symbols for RC files too.

Also increment the version as this change modifies the output for the existing
bakefiles.