Implement exclusion of sources and targets in VS backends.
Support generating Visual Studio projects with files excluded from the
build in some configurations.  Also implement exclusion of whole
targets/projects from the solution.

To use, just enclose the sources {} statement or the target inside an if
block that tests the value of $(config).

Updated tests/projects/conditional to demonstrate and test this.