Don't pass -pthread to the linker on OS X.
The -pthread option is noop with GCC on Darwin, because the necessary
runtime is included in libSystem; GCC silently accepts it for
cross-platform compatibility.

Unfortunately, Clang does not and issues a warning:
clang: warning: argument unused during compilation: '-pthread'

This is arguably a bug in Clang -- see
http://llvm.org/bugs/show_bug.cgi?id=7798. But since the bug was filed
almost three years ago and nobody got around to fixing it yet, it seems
prudent to just tweak the gnu-osx toolset accordingly. Fortunately, gcc
still works just fine without -pthread, so the fix is simple enough.