Add --diff-only option to only see changes to the generated files.
This is mostly useful when developing bakefile itself but could be also used
to check if the changes really have the expected effect in normal use too.

The diff is currently in simple unified format with default context and no
colours (but https://pypi.python.org/pypi/cdiff can be used with it).