Allow using settings with MSVS backends.
Currently the setting value is simply replaced with its default value and
can't be customized in any way, we should generate .props files allowing to do
it in the future. But even being able to use the default settings values is
useful as they allow injecting environment variables in the generated files.