Fix too eager expansion of variables inside "sources" statement.
Keep variables in the names of the sources unexpanded, otherwise they could be
evaluated too early, e.g. in the context of the base template where variables
had different values from their real values in the most derived target.

This was inconsistent with the use of variables in all the other contexts and
not very useful as it didn't allow the final target to specify where its
sources should be found, so fix this by not performing this early expansion.