Avoid spurious "unused variable" warnings for some settings.
A special artificial variable is created for each setting to make it possible
to reference it easier, however this variable itself is only referenced from a
PlaceholderExpr which wasn't checked for variable references before, resulting
in "unused variable" warnings being wrongly given.

Fix this by handling PlaceholderExpr correctly in _UsedVariablesTracker.