[gnu] Don't rely on "$_" being set in the generated makefiles.
It turns out that "$_" is not set correctly (at least if bash and zsh
behaviour is considered correct) in dash when using compound commands, which
means that using the build directory doesn't work on default Debian systems
that use dash as /bin/sh.

Separate setting the variable and calling mkdir to fix this and also add
better error checking to stop make immediately if we failed to create the
build directory.