[gnu] Get rid of "_builddir" in the intermediate targets names.
While this didn't seem to have any actually bad effects in practice, it seemed
wrong to put "$(_builddir)" in the names of intermediate targets for the
custom rules with multiple outputs, as we did, as it really has nothing to do
there.

At the very least this makes the generated makefiles slightly shorter.