Add "max" warnings level and restore "all" to mean /W4 for MSVS
Partially revert the changes of 7716c419b1617a360c3d38ffd1ab41bccd0bcd61 and
map "warnings = all" to "/W4" in MSVS projects as "EnableAllWarnings" is
unusable in practice, generating hundreds of warnings in the standard headers.

Still allow using it if the new "max" warnings level is specified and update
the other compilers to try to generate as many warnings as possible for this
level.