Allow MSVS plugin to avoid collapsing some tags in 201x format too
Move support for elems_not_collapsed down to the base XmlFormatter class from
VS200xXmlFormatter, so that it could be reused from VS 201x code too.

This also allows to avoid duplicating format_node() definition.

There are no changes to any output formats in this commit, this is a pure
refactoring.