Handle targets with toolset-specific external dependencies
If a target has a conditional external project as a dependency, with a
condition involving "$toolset" (e.g. in the simplest case just something like
"$toolset == vs2015"), we may end up with a project whose dependencies
evaluate to nothing for the toolset currently being generated and this
resulted in the rather unclear errors about "'NoneType' object is not
iterable".

Handle this gracefully instead by just ignoring the dependencies in this case.