Handle external MSBuild projects without configurations/platforms
Such projects do have configurations/platforms defined, of course, as any
valid project must, but we could well have failed to extract the information
about them from the project file if it defines them in another file which it
imports.

Currently this resulted in a crash due to trying to access the first element
of empty configurations/platforms lists and while the crash could be avoided,
the behaviour still wouldn't be very useful, so just assume that the project
supports all solution configurations if we don't know the project
configurations exactly.

This is clearly wrong and may result in build errors, but still more useful
than the old behaviour.