Recognize MSVS 12 tools version for external projects too
According to https://msdn.microsoft.com/en-us/library/bb383796.aspx, this is
the native value used by MSVS 2013, but we didn't recognize it before,
presumably because we only tested with the external projects originally
created in a previous MSVS version and updated to VS 2013 later.

Do recognize 12, in addition to 4 (MSVS 2010 and 2012) and 14 (MSVS 2015) too
now.