Update to PyInstaller 3.2
The previously used version of PyInstaller died with mysterious errors and,
instead of debugging them, just update to the latest version which seems to
work fine out of the box.

Also change the bkl.exe location, it is (now?) generated in a subdirectory.
And make this location a variable to make it simpler to change it later.