Switch to using parametrized instead of yielding tests
Pytest 3 deprecates yield tests, resulting in 83 warnings of the form

yield tests are deprecated, and scheduled to be removed in pytest 4.0

being given when running the test suite using it.

Replace them with the parametrized tests, as recommended.

There should be no actual changes to the tests behaviour, the same 91
tests still pass before and after this change.