Make values of the "deps" property unique
Add "idempotent" attribute for the properties which ignores appending an
already existing item to a property of list type, making it set-like,
and set this attribute to true for the "deps" property to avoid having
duplicate dependencies in it, which resulted in wrong MSVS toolsets
output as the projects dependencies are supposed to be unique.

Another way of achieving the same goal would be to add a SetType, in
addition to the existing ListType, but it seems more complicated and
it's not really clear if it's worth it (in particular, efficiency is
probably not a concern here).