Add a test checking that dependencies are not duplicated
Although nobody would arguably duplicate the dependencies on purpose, it
was easy to do this unintentionally by inheriting a dependency from a
base template and also adding it explicitly -- and this, in fact,
happened in several places in real projects.

Add a unit test verifying that this now behaves as expected and doesn't
generate duplicate dependencies in the output solution file.