Allow explicitly specifying configs/platforms for externals
Auto-detecting configurations and platforms used by external MSVS files
doesn't always work, e.g. it fails to detect them for C++ REST SDK
projects because they define them in a separate MSBuild file included
from the main project file, and so not seen in its XML.

Allow working around this by respecting the explicitly defined "archs"
and "configurations" properties for an external -- now it's possible to
do it to fix multiple problems in the generated files for MSVS toolsets
due to assuming that the external supported all configurations/platforms
used by the solution.