makefile_defs_mingw.bkl: The object file list is written to an .in file in linking statements which the linker reads in upon invocation (using the '@file' syntax).
Backslashes in the object file paths are converted to forward slashes
to avoid issues with compilers.

If make is run with SHELL = cmd, the make rule command is written to a
temporary batch file which is then called to execute the command. However,
the length of each line of a batch file is subject to the maximum length
of a cmd command string.

The approach below should work for every shell (incl. unix-like shells).

The .in file is intentionally not removed after command execution.