Make generation of solution file in VSSolutionBase unconditional
Instead of testing whether the .sln file should be generated in
VSSolutionBase.write() itself, create a different VSSolutionNull object
if we don't need to generate the solution file at all.

This is a minor optimization, as we avoid doing some unnecessary work,
and will be simpler to generalize for the case when more than one
solution file may need to be generated in the upcoming commits.