Fix bug with using generate-solution=no in a submodule
In this case the main MSVSSolutionsBundle object tried to check if a
solution needed to be generated for the submodule by checking its
"solutions" attribute, but this attribute didn't exist in VSSolutionNull
which was created for the submodule if solutions were not generated for
it at all.

Fix this by simply adding a dummy solutions attribute to VSSolutionNull,
it's a bit ugly but does no real harm.