Fix generation of MSVS projects when solutions are not generated
This logically reverts 8e4c7dc7 (Make generation of solution file in
VSSolutionBase unconditional, 2020-09-14) which added VSSolutionNull
replacing explicit tests for the solution generation: this looked like a
good idea and did simplify the code a bit, but didn't actually work,
because all the projects must be part of the global project tree for
functions such as _find_target_guid_recursively and almost certainly
others to work, as they need to recurse inside the subsolutions field
which didn't exist in VSSolutionNull.

It would have been also possible to add subsolutions (and also projects)
members to VSSolutionNull, but this wouldn't make it obviously simpler
than the version about it any more, so just remove it instead.